refactor(rules): Migrate shared consumers to ps.Resolver
Switch the rule engine, compiler, sequence state, and filter options to the narrow Resolver interface, and add Linux filter/action stubs so rules compile and evaluate on Linux.
