+++ pkg/filter/accessor.go
@@ -152,87 +152,7 @@ func (*evtAccessor) Get(f Field, evt *event.Event) (params.Value, error) {
 // as a regular LHS/RHS component, used as a function parameter or
 // referenced in the bound field.
 func (f *filter) narrowAccessors() {
-	var (
-		removeEvtAccessor        = true
-		removePsAccessor         = true
-		removeThreadAccessor     = true
-		removeModuleAccessor     = true
-		removeFileAccessor       = true
-		removeRegistryAccessor   = true
-		removeNetworkAccessor    = true
-		removeHandleAccessor     = true
-		removePEAccessor         = true
-		removeMemAccessor        = true
-		removeDNSAccessor        = true
-		removeThreadpoolAccessor = true
-	)
-
-	for _, field := range f.fields {
-		switch {
-		case field.Name.IsEvtField() || field.Name.IsKevtField():
-			removeEvtAccessor = false
-		case field.Name.IsPeField():
-			removePEAccessor = false
-		case field.Name.IsPsField():
-			removePsAccessor = false
-		case field.Name.IsThreadField():
-			removeThreadAccessor = false
-		case field.Name.IsImageField() || field.Name.IsModuleField():
-			removeModuleAccessor = false
-		case field.Name.IsFileField():
-			removeFileAccessor = false
-		case field.Name.IsRegistryField():
-			removeRegistryAccessor = false
-		case field.Name.IsNetworkField():
-			removeNetworkAccessor = false
-		case field.Name.IsHandleField():
-			removeHandleAccessor = false
-		case field.Name.IsMemField():
-			removeMemAccessor = false
-		case field.Name.IsDNSField():
-			removeDNSAccessor = false
-		case field.Name.IsThreadpoolField():
-			removeThreadpoolAccessor = false
-		}
-	}
-
-	if removeEvtAccessor {
-		f.removeAccessor(&evtAccessor{})
-	}
-	if removePsAccessor {
-		f.removeAccessor(&psAccessor{})
-	}
-	if removeThreadAccessor {
-		f.removeAccessor(&threadAccessor{})
-	}
-	if removeModuleAccessor {
-		f.removeAccessor(&moduleAccessor{})
-	}
-	if removeFileAccessor {
-		f.removeAccessor(&fileAccessor{})
-	}
-	if removeRegistryAccessor {
-		f.removeAccessor(&registryAccessor{})
-	}
-	if removeNetworkAccessor {
-		f.removeAccessor(&networkAccessor{})
-	}
-	if removeHandleAccessor {
-		f.removeAccessor(&handleAccessor{})
-	}
-	if removePEAccessor {
-		f.removeAccessor(&peAccessor{})
-	}
-	if removeMemAccessor {
-		f.removeAccessor(&memAccessor{})
-	}
-	if removeDNSAccessor {
-		f.removeAccessor(&dnsAccessor{})
-	}
-	if removeThreadpoolAccessor {
-		f.removeAccessor(&threadpoolAccessor{})
-	}
-
+	f.pruneUnusedAccessors()
 	for _, accessor := range f.accessors {
 		accessor.SetFields(f.fields)
 		accessor.SetSegments(f.segments)
