+++ pkg/filter/ql/functions/minidump_linux.go
@@ -1,0 +1,39 @@
+//go:build linux
+
+/*
+ * Copyright 2026 by Nedim Sabic Sabic
+ * https://www.fibratus.io
+ * All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package functions
+
+// IsMinidump determines if the specified file contains the minidump signature.
+type IsMinidump struct{}
+
+func (f IsMinidump) Call(args []interface{}) (interface{}, bool) {
+	return false, len(args) >= 1
+}
+
+func (f IsMinidump) Desc() FunctionDesc {
+	return FunctionDesc{
+		Name: IsMinidumpFn,
+		Args: []FunctionArgDesc{
+			{Keyword: "file", Types: []ArgType{Field, BoundField, BoundSegment, BareBoundVariable, Func, String}, Required: true},
+		},
+	}
+}
+
+func (f IsMinidump) Name() Fn { return IsMinidumpFn }
