+++ pkg/rules/action/isolate_linux.go
@@ -1,0 +1,32 @@
+//go:build linux
+
+/*
+ * Copyright 2026 by Nedim Sabic Sabic
+ * https://www.fibratus.io
+ * All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package action
+
+import (
+	"fmt"
+	"net"
+)
+
+// Isolate isolates the host via firewall rules.
+// Linux isolation is not wired yet.
+func Isolate(whitelist []net.IP) error {
+	return fmt.Errorf("isolate action is not implemented on Linux yet (whitelist=%v)", whitelist)
+}
