+++ pkg/rules/action/kill_linux.go
@@ -1,0 +1,29 @@
+//go:build linux
+
+/*
+ * Copyright 2026 by Nedim Sabic Sabic
+ * https://www.fibratus.io
+ * All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package action
+
+import "fmt"
+
+// Kill terminates all processes with specified pids.
+// The signal-based Linux implementation lands with the detection work.
+func Kill(pids []uint32) error {
+	return fmt.Errorf("kill action is not implemented on Linux yet (pids=%v)", pids)
+}
