+++ pkg/rules/compiler.go
@@ -58,12 +58,12 @@ var (
 )
 
 type compiler struct {
-	psnap     ps.Snapshotter
+	psnap     ps.Resolver
 	config    *config.Config
 	approvers config.Approvers
 }
 
-func newCompiler(psnap ps.Snapshotter, cfg *config.Config) *compiler {
+func newCompiler(psnap ps.Resolver, cfg *config.Config) *compiler {
 	return &compiler{psnap: psnap, config: cfg, approvers: config.Approvers{
 		Keys:        make(map[string][]string),
 		Paths:       make(map[string][]string),
