+++ pkg/rules/engine_linux_test.go
@@ -1,0 +1,36 @@
+//go:build linux
+
+/*
+ * Copyright 2026 by Nedim Sabic Sabic
+ * https://www.fibratus.io
+ * All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package rules
+
+import (
+	"testing"
+
+	"github.com/rabbitstack/fibratus/pkg/config"
+	"github.com/rabbitstack/fibratus/pkg/ps"
+	"github.com/stretchr/testify/require"
+)
+
+func TestNewEngineAcceptsResolver(t *testing.T) {
+	cfg := &config.Config{Filters: &config.Filters{}}
+	e := NewEngine(ps.NewLinuxSnapshotter(), cfg)
+	require.NotNil(t, e)
+	require.NotNil(t, e.compiler)
+}
